+++ docs/src/news.md
@@ -1,6 +1,10 @@
 ### NEWS for package **Jchemo**
 
 ### *Version 0.1.13* 
+- News
+    - **mtest** Select indexes defining training and test sets for each column 
+        of a dataframe
+
 - Modifications
     - Code cleaning